pyproject.toml
[project]
name = "fire-detection-iot"
version = "0.1.0"
description = ""
authors = [
    {name = "Ganga-Suren",email = "basvag34@students.rowan.edu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "boto3 (>=1.38.11,<2.0.0)",
    "datetime (>=5.5,<6.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "fire_detection_iot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


